+++ sound-mind-studio/tests/test_macro_recorder.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <QObject>
+
+class MacroRecorderTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void freshRecorderIsNotRecordingAndHasNoEvents();
+    void startRecordingBeginsRecordingAndClearsPreviousEvents();
+    void recordEventIsANoOpWhileNotRecording();
+    void recordEventAppendsWhileRecording();
+    void stopRecordingStopsButKeepsEventsAvailable();
+    void startRecordingWhileAlreadyRecordingDoesNotClearEvents();
+    void stopRecordingWhileNotRecordingIsANoOp();
+    void recordingStateChangedEmitsOnStartAndStop();
+};